# docs/openapi_spec.json
{
  "definitions": {
...
    "LoginRequest": {
      "properties": {
        "identifier": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "description": "Username, email, or employee ID (for password login)",
          "title": "Identifier"
        },
        "mobile": {
          "anyOf": [
            {
              "maxLength": 15,
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "description": "Mobile number (for OTP login)",
          "title": "Mobile"
        },
        "otp": {
          "anyOf": [
            {
              "maxLength": 6,
              "minLength": 4,
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "description": "OTP (for OTP login)",
          "title": "Otp"
        },
        "password": {
          "anyOf": [
            {
              "minLength": 8,
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "description": "Password (for password login)",
          "title": "Password"
        }
      },
      "title": "LoginRequest",
      "type": "object"
    },
...
    "UserCreateSchema": {
      "additionalProperties": false,
      "properties": {
        "_id": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Id"
        },
        "created_at": {
          "anyOf": [
            {
              "format": "date-time",
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Created At"
        },
        "deleted_at": {
          "anyOf": [
            {
              "format": "date-time",
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Deleted At"
        },
        "department": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Department"
        },
        "email": {
          "anyOf": [
            {
              "pattern": "^\\S+@\\S+\\.\\S+$",
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Email"
        },
        "employee_id": {
          "anyOf": [
            {
              "maxLength": 30,
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Employee Id"
        },
        "failed_login_attempts": {
          "default": 0,
          "title": "Failed Login Attempts",
          "type": "integer"
        },
        "is_active": {
          "default": true,
          "title": "Is Active",
          "type": "boolean"
        },
        "is_admin": {
          "default": false,
          "title": "Is Admin",
          "type": "boolean"
        },
        "is_deleted": {
          "default": false,
          "title": "Is Deleted",
          "type": "boolean"
        },
        "is_email_verified": {
          "default": false,
          "title": "Is Email Verified",
          "type": "boolean"
        },
        "last_login": {
          "anyOf": [
            {
              "format": "date-time",
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Last Login"
        },
        "lock_until": {
          "anyOf": [
            {
              "format": "date-time",
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Lock Until"
        },
        "mobile": {
          "anyOf": [
            {
              "maxLength": 15,
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Mobile"
        },
        "organization_id": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Organization Id"
        },
        "otp_resend_count": {
          "default": 0,
          "title": "Otp Resend Count",
          "type": "integer"
        },
        "password": {
          "minLength": 8,
          "title": "Password",
          "type": "string"
        },
        "roles": {
          "items": {
            "enum": [
              "superadmin",
              "admin",
              "user",
              "creator",
              "editor",
              "publisher",
              "deo",
              "manager",
              "general"
            ],
            "type": "string"
          },
          "title": "Roles",
          "type": "array"
        },
        "updated_at": {
          "anyOf": [
            {
              "format": "date-time",
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Updated At"
        },
        "user_type": {
          "enum": [
            "employee",
            "general"
          ],
          "title": "User Type",
          "type": "string"
        },
        "username": {
          "anyOf": [
            {
              "maxLength": 50,
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Username"
        }
      },
      "required": [
        "user_type",
        "password"
      ],
      "title": "UserCreateSchema",
      "type": "object"
    },
...
    "UserOut": {
      "description": "Schema for public API responses. \nExcludes internal security counters.",
      "properties": {
        "created_at": {
          "anyOf": [
            {
              "format": "date-time",
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Created At"
        },
        "deleted_at": {
          "anyOf": [
            {
              "format": "date-time",
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Deleted At"
        },
        "department": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Department"
        },
        "email": {
          "anyOf": [
            {
              "pattern": "^\\S+@\\S+\\.\\S+$",
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Email"
        },
        "employee_id": {
          "anyOf": [
            {
              "maxLength": 30,
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Employee Id"
        },
        "failed_login_attempts": {
          "default": 0,
          "title": "Failed Login Attempts",
          "type": "integer"
        },
        "id": {
          "title": "Id",
          "type": "string"
        },
        "is_active": {
          "default": true,
          "title": "Is Active",
          "type": "boolean"
        },
        "is_admin": {
          "default": false,
          "title": "Is Admin",
          "type": "boolean"
        },
        "is_deleted": {
          "default": false,
          "title": "Is Deleted",
          "type": "boolean"
        },
        "is_email_verified": {
          "default": false,
          "title": "Is Email Verified",
          "type": "boolean"
        },
        "last_login": {
          "anyOf": [
            {
              "format": "date-time",
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Last Login"
        },
        "lock_until": {
          "anyOf": [
            {
              "format": "date-time",
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Lock Until"
        },
        "mobile": {
          "anyOf": [
            {
              "maxLength": 15,
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Mobile"
        },
        "organization_id": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Organization Id"
        },
        "otp_resend_count": {
          "default": 0,
          "title": "Otp Resend Count",
          "type": "integer"
        },
        "roles": {
          "items": {
            "enum": [
              "superadmin",
              "admin",
              "user",
              "creator",
              "editor",
              "publisher",
              "deo",
              "manager",
              "general"
            ],
            "type": "string"
          },
          "title": "Roles",
          "type": "array"
        },
        "updated_at": {
          "anyOf": [
            {
              "format": "date-time",
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Updated At"
        },
        "user_type": {
          "enum": [
            "employee",
            "general"
          ],
          "title": "User Type",
          "type": "string"
        },
        "username": {
          "anyOf": [
            {
              "maxLength": 50,
              "type": "string"
            },
            {
              "type": "null"
            }
          ],
          "default": null,
          "title": "Username"
        }
      },
      "required": [
        "id",
        "user_type"
      ],
      "title": "UserOut",
      "type": "object"
    },
...
  },
  "info": {
    "description": "Comprehensive API for managing forms, responses, and AI analysis.",
    "title": "Forms Backend API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/form/api/v1/auth/login": {
      "post": {
        "description": "Authenticate via password or OTP and issue JWT tokens.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "properties": {
                "data": {
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "refresh_token": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/definitions/UserOut"
                    }
                  },
                  "type": "object"
                },
                "success": {
                  "type": "boolean"
                }
              },
              "type": "object"
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Authenticate via password or OTP and issue JWT tokens.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/form/api/v1/auth/logout": {
      "post": {
        "description": "Revokes the user's access and refresh tokens.",
        "responses": {
          "200": {
            "description": "Logout successful"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Revoke the current JWT session.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/form/api/v1/auth/refresh": {
      "post": {
        "description": "Generates a new access token using a valid refresh token.",
        "responses": {
          "200": {
            "description": "Token refreshed",
            "schema": {
              "properties": {
                "access_token": {
                  "type": "string"
                },
                "refresh_token": {
                  "type": "string"
                },
                "success": {
                  "type": "boolean"
                }
              },
              "type": "object"
            }
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Issue a new access token using a valid refresh token.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/form/api/v1/auth/register": {
      "post": {
        "description": "Registers a new user and returns user details.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserCreateSchema"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully",
            "schema": {
              "$ref": "#/definitions/UserOut"
            }
          },
          "400": {
            "description": "Validation error or user already exists"
          }
        },
        "summary": "Register a new user account.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/form/api/v1/auth/request-otp": {
      "post": {
        "description": "Generate and send an OTP to the given mobile/email.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "email": {
                  "type": "string"
                },
                "mobile": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OTP sent successfully"
          },
          "400": {
            "description": "Missing identifier"
          }
        },
        "summary": "Generate and send an OTP to the given mobile/email.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/form/api/v1/auth/revoke-all": {
      "post": {
        "description": "Revokes all active JWT sessions for the authenticated user.",
        "responses": {
          "200": {
            "description": "All sessions revoked successfully"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Revoke all active sessions for the authenticated user.",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}